requirements.txt
streamlit
google-genai
xhtml2pdf
markdown